app/requirements.txt
openai-whisper>=20231117
librosa>=0.10.0